+++ src/api/dto/character-overall-res.dto.ts
@@ -1,0 +1,3 @@
+import { CharacterDto } from 'src/common/dto/character.dto';
+
+export class CharacterOverallResDto extends CharacterDto {}
