--- src/api/dto/get-character-overall.dto.ts
@@ -1,9 +1,0 @@
-import { Expose } from 'class-transformer';
-import { IsBoolean, IsOptional } from 'class-validator';
-
-export class GetCharacterOverallDto {
-  @Expose()
-  @IsOptional()
-  @IsBoolean()
-  update?: boolean;
-}
