+++ src/character/repository/character.repository.ts
@@ -1,16 +1,15 @@
 import { Injectable } from '@nestjs/common';
-import { removeNulls } from 'src/util/removeNulls';
+import { CharacterDto } from 'src/common/dto/character.dto';
 import { PrismaService } from '../../prisma/prisma.service';
-import { AbilityInfo, CharacterAbility } from '../type/character-ability.type';
-import { HyperStatInfo } from '../type/character-hyper-stat.type';
-import { ItemEquipmentInfo, ItemOption } from '../type/character-item-equipment.type';
-import { Character } from '../type/character.type';
+import { convertAbilityToDto, convertAbilityToEntity } from '../converter/ability.converter';
+import { convertHyperStatToDto, convertHyperStatToEntity } from '../converter/hyper-stat.converter';
+import { convertItemEquipmentToDto, convertItemEquipmentToEntity } from '../converter/item-equipment.converter';
 
 @Injectable()
 export class CharacterRepository {
   constructor(private readonly prismaService: PrismaService) {}
 
-  async findCharacterOverallByNickname(nickname: string): Promise<Character> {
+  async findCharacterOverallByNickname(nickname: string): Promise<CharacterDto> {
     const characterData = await this.prismaService.character.findUnique({
       where: {
         nickname,
@@ -51,385 +50,38 @@ export class CharacterRepository {
 
     if (!characterData) return null;
 
-    const hyperStatPresets = [
-      { presetNo: 1, active: false, remainPoint: 0, hyperStatInfo: [] },
-      { presetNo: 2, active: false, remainPoint: 0, hyperStatInfo: [] },
-      { presetNo: 3, active: false, remainPoint: 0, hyperStatInfo: [] },
-    ];
-
-    const abilityPresets: CharacterAbility = {
-      remainFame: 0, // preset 0번의 abilityNo 값
-      preset: [
-        { presetNo: 1, active: false, abilityInfo: [] },
-        { presetNo: 2, active: false, abilityInfo: [] },
-        { presetNo: 3, active: false, abilityInfo: [] },
-      ],
-    };
-
-    const itemEquipmentPresets = [
-      { presetNo: 1, active: false, itemEquipmentInfo: [] },
-      { presetNo: 2, active: false, itemEquipmentInfo: [] },
-      { presetNo: 3, active: false, itemEquipmentInfo: [] },
-      { presetNo: 4, active: false, itemEquipmentInfo: [] }, // 타이틀
-      { presetNo: 5, active: false, itemEquipmentInfo: [] }, // 드래곤 or 메카닉 장비
-    ];
-
-    // 2. 하이퍼스탯 데이터를 프리셋별로 그룹화
-    if (characterData.hyperStatPreset) {
-      characterData.hyperStatPreset.forEach((hs) => {
-        const preset = hyperStatPresets.find((p) => p.presetNo === hs.presetNo);
-        if (preset) {
-          preset.active = hs.active; // 활성화 여부 설정
-          if (hs.hyperStat.statType === '_REMAIN_POINT') {
-            preset.remainPoint = hs.hyperStat.statPoint;
-          } else {
-            preset.hyperStatInfo.push(hs.hyperStat);
-          }
-        }
-      });
-    }
-
-    // 3. 어빌리티 데이터를 프리셋별로 그룹화
-    if (characterData.abilityPreset) {
-      characterData.abilityPreset.forEach((ab) => {
-        if (ab.presetNo === 0) {
-          abilityPresets.remainFame = ab.abilityNo;
-        }
-
-        const preset = abilityPresets.preset.find((p) => p.presetNo === ab.presetNo);
-        if (preset) {
-          preset.active = ab.active; // 활성화 여부 설정
-          preset.abilityInfo.push({
-            abilityNo: ab.abilityNo,
-            abilityGrade: ab.ability.abilityGrade,
-            abilityValue: ab.ability.abilityValue,
-          });
-        }
-      });
-    }
-
-    // 4. 장비 데이터를 프리셋별로 그룹화
-    const getPotentialText = (itemEquipment, potentialNo: number) => {
-      return itemEquipment.ItemEquipmentPotential.find((potential) => potential.potentialNo === potentialNo)?.potential
-        ?.potential;
-    };
-
-    const getPotentials = (itemEquipment) => {
-      const res = {};
-      if (itemEquipment.potentialOptionGrade) {
-        res['potentialOption'] = [
-          getPotentialText(itemEquipment, 1),
-          getPotentialText(itemEquipment, 2),
-          getPotentialText(itemEquipment, 3),
-        ];
-      }
-      if (itemEquipment.additionalPotentialOptionGrade) {
-        res['additionalPotentialOption'] = [
-          getPotentialText(itemEquipment, 4),
-          getPotentialText(itemEquipment, 5),
-          getPotentialText(itemEquipment, 6),
-        ];
-      }
-      return res;
-    };
-
-    const getOption = (itemEquipment, optionType: 'TOTAL' | 'BASE' | 'EXCEPTIONAL' | 'ADD' | 'ETC' | 'STARFORCE') => {
-      return itemEquipment.ItemEquipmentOption.find((option) => option.optionType === optionType)?.option;
-    };
-
-    if (characterData.itemEquipmentPreset) {
-      characterData.itemEquipmentPreset.forEach((eq) => {
-        const preset = itemEquipmentPresets.find((p) => p.presetNo === eq.presetNo);
-        if (preset) {
-          preset.active = eq.active; // 활성화 여부 설정
-          preset.itemEquipmentInfo.push({
-            ...eq.itemEquipment,
-            ...getPotentials(eq.itemEquipment),
-            totalOption: getOption(eq.itemEquipment, 'TOTAL'),
-            baseOption: getOption(eq.itemEquipment, 'BASE'),
-            exceptionalOption: getOption(eq.itemEquipment, 'EXCEPTIONAL'),
-            addOption: getOption(eq.itemEquipment, 'ADD'),
-            etcOption: getOption(eq.itemEquipment, 'ETC'),
-            starforceOption: getOption(eq.itemEquipment, 'STARFORCE'),
-          });
-        }
-      });
-    }
-
     //? 원래도 이거 안빼면 upsert 오류났었나..?
     const { id, statId, propensityId, abilityPreset, ...characterBasic } = characterData;
 
-    const res = {
+    return {
       ...characterBasic,
-      hyperStatPreset: characterData.hyperStatPreset ? hyperStatPresets : null,
-      ability: characterData.abilityPreset ? abilityPresets : null,
-      itemEquipmentPreset: characterData.itemEquipmentPreset ? itemEquipmentPresets : null,
+      hyperStatPreset: convertHyperStatToDto(characterData.hyperStatPreset),
+      ability: convertAbilityToDto(characterData.abilityPreset),
+      itemEquipmentPreset: convertItemEquipmentToDto(characterData.itemEquipmentPreset),
     };
-
-    return removeNulls(res);
   }
 
-  async createOrIgnoreHyperstat(hyperStats: HyperStatInfo[]) {
-    // ON CONFLICT 시 auto_increment가 너무 많이 튀는거 방지용
-    // 캐릭터가 많아질수록 insert 할 행이 줄고, 경쟁조건이 많이 없을것으로 예상해서 이정도로 처리
-
-    // 들어온 값에서 중복유무 체크
-    const uniqueHyperStats = Array.from(
-      new Map(hyperStats.map((hs) => [hs.statType + '-' + hs.statPoint, hs])).values(),
-    );
-
-    const existingHyperStats = await this.prismaService.hyperStat.findMany({
-      where: {
-        OR: uniqueHyperStats.map((hs) => ({
-          statType: hs.statType,
-          statPoint: hs.statPoint,
-        })),
-      },
-      select: { statType: true, statPoint: true },
-    });
-
-    const newHyperStats = uniqueHyperStats.filter(
-      (hs) => !existingHyperStats.some((ehs) => ehs.statType === hs.statType && ehs.statPoint === hs.statPoint),
-    );
-    return this.prismaService.hyperStat.createMany({
-      data: newHyperStats,
-      skipDuplicates: true,
-    });
-  }
-
-  async createOrIgnoreAbility(ability: AbilityInfo[]) {
-    const uniqueAbilities = Array.from(new Map(ability.map((ab) => [ab.abilityValue, ab])).values());
-
-    const existingAbilities = await this.prismaService.ability.findMany({
-      where: {
-        abilityValue: { in: uniqueAbilities.map((ab) => ab.abilityValue) },
-      },
-      select: { abilityValue: true },
-    });
-
-    const newAbilities = uniqueAbilities
-      .filter((ab) => !existingAbilities.some((eab) => eab.abilityValue === ab.abilityValue))
-      .map(({ abilityNo, ...rest }) => rest);
-    return this.prismaService.ability.createMany({
-      data: newAbilities,
-      skipDuplicates: true,
-    });
-  }
-
-  private async createOrIgnoreItemOption(itemOptions: ItemOption[]) {
-    const uniqueItemOptions = Array.from(new Map(itemOptions.map((io) => [io.hash, io])).values());
-
-    const existingItemOptions = await this.prismaService.itemOption.findMany({
-      where: {
-        hash: { in: uniqueItemOptions.map((io) => io.hash) },
-      },
-      select: { hash: true },
-    });
-
-    const newOptions = uniqueItemOptions.filter((io) => !existingItemOptions.some((eio) => eio.hash === io.hash));
-    return this.prismaService.itemOption.createMany({
-      data: newOptions,
-      skipDuplicates: true,
-    });
-  }
-
-  private async createOrIgnoreItemPotential(itemPotentials: string[]) {
-    const uniqueItemPotentials = Array.from(new Set(itemPotentials));
-
-    const existingItemPotentials = await this.prismaService.potential.findMany({
-      where: {
-        potential: { in: uniqueItemPotentials },
-      },
-      select: { potential: true },
-    });
-
-    const newPotentials = uniqueItemPotentials
-      .filter((potential) => !existingItemPotentials.some((eip) => eip.potential === potential))
-      .map((potential) => ({ potential }));
-    return this.prismaService.potential.createMany({
-      data: newPotentials,
-      skipDuplicates: true,
-    });
-  }
-
-  private createOptionConnections = (itemEquipment: ItemEquipmentInfo, optionIdMap: Record<string, number>) => {
-    const optionTypes = [
-      { option: itemEquipment.totalOption, type: 'TOTAL' },
-      { option: itemEquipment.baseOption, type: 'BASE' },
-      { option: itemEquipment.exceptionalOption, type: 'EXCEPTIONAL' },
-      { option: itemEquipment.addOption, type: 'ADD' },
-      { option: itemEquipment.etcOption, type: 'ETC' },
-      { option: itemEquipment.starforceOption, type: 'STARFORCE' },
-    ];
-
-    // 각 옵션별 연결 가능하도록 생성
-    return optionTypes
-      .filter(({ option }) => option && optionIdMap[option.hash])
-      .map(({ option, type }) => ({
-        option: {
-          connect: {
-            id: optionIdMap[option.hash],
-          },
-        },
-        optionType: type as 'TOTAL' | 'BASE' | 'EXCEPTIONAL' | 'ADD' | 'ETC' | 'STARFORCE',
-      }));
-  };
-
-  private createPotentialConnections = (itemEquipment: ItemEquipmentInfo, potentialIdMap: Record<string, number>) => {
-    const potentialOptions = itemEquipment.potentialOption || [];
-    const additionalPotentialOptions = itemEquipment.additionalPotentialOption || [];
-
-    const allPotentials = [
-      ...potentialOptions.map((potential, index) => ({ potential, potentialNo: index + 1 })),
-      ...additionalPotentialOptions.map((potential, index) => ({ potential, potentialNo: index + 4 })),
-    ];
-
-    return allPotentials
-      .filter(({ potential }) => potentialIdMap[potential])
-      .map(({ potential, potentialNo }) => ({
-        potential: {
-          connect: {
-            id: potentialIdMap[potential],
-          },
-        },
-        potentialNo,
-      }));
-  };
-
-  async createOrIgnoreItemEquipment(itemEquipment: ItemEquipmentInfo[]) {
-    // 장비마다의 모든 옵션에 대해 저장
-    const flatOption = itemEquipment.flatMap((item) => {
-      return [
-        item?.totalOption,
-        item?.baseOption,
-        item?.exceptionalOption,
-        item?.addOption,
-        item?.etcOption,
-        item?.starforceOption,
-      ].filter(Boolean);
-    });
-    await this.createOrIgnoreItemOption(flatOption);
-
-    // 잠재능력
-    const flatPotential = itemEquipment.flatMap((item) => {
-      return [...(item?.potentialOption ?? []), ...(item?.additionalPotentialOption ?? [])].filter(Boolean);
-    });
-    await this.createOrIgnoreItemPotential(flatPotential);
-
-    // 장비 저장
-    for (const itemEquip of itemEquipment) {
-      // 장비 해시 존재 여부 확인
-      const existingItemEquipment = await this.prismaService.itemEquipment.findUnique({
-        where: { hash: itemEquip.hash },
-      });
-
-      // 이미 존재하는 장비라면 넘어감
-      if (existingItemEquipment) continue;
-
-      // 장비 옵션 해시만 필터링
-      const optionHashes = [
-        itemEquip.totalOption?.hash,
-        itemEquip.baseOption?.hash,
-        itemEquip.exceptionalOption?.hash,
-        itemEquip.addOption?.hash,
-        itemEquip.etcOption?.hash,
-        itemEquip.starforceOption?.hash,
-      ].filter(Boolean);
-
-      const optionIds = await this.prismaService.itemOption.findMany({
-        where: { hash: { in: optionHashes } },
-        select: { id: true, hash: true },
-      });
-
-      // 잠재능력 해시만 필터링
-      const potentialHashes = [...itemEquip.potentialOption, ...itemEquip.additionalPotentialOption].filter(Boolean);
-
-      const potentialIds = await this.prismaService.potential.findMany({
-        where: { potential: { in: potentialHashes } },
-        select: { id: true, potential: true },
-      });
-
-      // 옵션, 잠재 해시랑 ID 매핑
-      const optionIdMap = Object.fromEntries(optionIds.map((option) => [option.hash, option.id]));
-      const potentialIdMap = Object.fromEntries(potentialIds.map((potential) => [potential.potential, potential.id]));
-
-      // 연결 생성준비
-      const optionConnections = this.createOptionConnections(itemEquip, optionIdMap);
-      const potentialConnections = this.createPotentialConnections(itemEquip, potentialIdMap);
-
-      // 장비 생성, 단 고유키 충돌로 실패시 무시
-
-      // 옵션, 잠재능력값 삭제
-      const {
-        totalOption,
-        baseOption,
-        exceptionalOption,
-        addOption,
-        etcOption,
-        starforceOption,
-        potentialOption,
-        additionalPotentialOption,
-        ...itemEquipWithoutOption
-      } = itemEquip;
-
-      try {
-        await this.prismaService.itemEquipment.create({
-          data: {
-            ...itemEquipWithoutOption,
-            ItemEquipmentOption: {
-              create: optionConnections,
-            },
-            ItemEquipmentPotential: {
-              create: potentialConnections,
-            },
-          },
-        });
-      } catch (e) {
-        if (e.code === 'P2002') continue;
-        throw e;
-      }
-    }
-  }
-
-  async upsertCharacterOverall(character: Character) {
+  async upsertCharacterOverall(character: CharacterDto) {
     const {
       stat,
-      hyperStatPreset: characterHyperStatList,
+      hyperStatPreset,
       propensity,
-      ability: characterAbilityList,
+      ability,
       itemEquipmentPreset: characterItemEquipmentList,
       ...characterBasic
     } = character;
 
     // 0. 하이퍼스탯, 어빌리티, 장비 데이터 flatten, 상위 데이터 적절히 삽입
-    const hyperStat = characterHyperStatList.flatMap((hs) => {
-      return hs.hyperStatInfo.map((h) => ({
-        ...h,
-        presetNo: hs.presetNo,
-        active: hs.active,
-      }));
-    });
-    const ability = characterAbilityList.preset.flatMap((ab) => {
-      return ab.abilityInfo.map((a) => ({
-        ...a,
-        presetNo: ab.presetNo,
-        active: ab.active,
-      }));
-    });
-    const itemEquipment = characterItemEquipmentList.flatMap((eq) => {
-      return eq.itemEquipmentInfo.map((e) => ({
-        ...e,
-        presetNo: eq.presetNo,
-        active: eq.active,
-      }));
-    });
+    const flatHyperStat = convertHyperStatToEntity(hyperStatPreset);
+    const flatAbility = convertAbilityToEntity(ability);
+    const flatItemEquipment = convertItemEquipmentToEntity(characterItemEquipmentList);
 
     // 1. 트랜잭션 밖에서 필요한 ID를 고유값을 사용해 한 번에 조회
     const [hyperStatIds, abilityIds, itemEquipmentIds] = await Promise.all([
       // HyperStat IDs 조회
       this.prismaService.hyperStat.findMany({
         where: {
-          OR: hyperStat.map((hs) => ({
+          OR: flatHyperStat.map((hs) => ({
             statType: hs.statType,
             statPoint: hs.statPoint,
           })),
@@ -440,7 +92,7 @@ export class CharacterRepository {
       // Ability IDs 조회
       this.prismaService.ability.findMany({
         where: {
-          abilityValue: { in: ability.map((ab) => ab.abilityValue) },
+          abilityValue: { in: flatAbility.map((ab) => ab.abilityValue) },
         },
         select: { id: true, abilityValue: true },
       }),
@@ -448,7 +100,7 @@ export class CharacterRepository {
       // ItemEquipment IDs 조회
       this.prismaService.itemEquipment.findMany({
         where: {
-          hash: { in: itemEquipment.map((eq) => eq.hash) },
+          hash: { in: flatItemEquipment.map((eq) => eq.hash) },
         },
         select: { id: true, hash: true },
       }),
@@ -499,7 +151,7 @@ export class CharacterRepository {
 
       // 하이퍼스탯 중간 테이블 연결
       await prisma.characterHyperStat.createMany({
-        data: hyperStat
+        data: flatHyperStat
           .filter((hs) => hyperStatMap.has(`${hs.statType}_${hs.statPoint}`)) // 존재하는 ID만 사용
           .map((hs) => ({
             characterId: characterId,
@@ -512,7 +164,7 @@ export class CharacterRepository {
 
       // 어빌리티 연결
       await prisma.characterAbility.createMany({
-        data: ability
+        data: flatAbility
           .filter((ab) => abilityMap.has(ab.abilityValue)) // 존재하는 ID만 사용
           .map((ab) => ({
             characterId: characterId,
@@ -526,7 +178,7 @@ export class CharacterRepository {
 
       // 장비 연결
       await prisma.characterItemEquipment.createMany({
-        data: itemEquipment
+        data: flatItemEquipment
           .filter((eq) => itemEquipmentMap.has(eq.hash)) // 존재하는 ID만 사용
           .map((eq) => ({
             characterId: characterId,
