--- src/character/type/character-propensity.type.ts
@@ -1,9 +1,0 @@
-export interface CharacterPropensity {
-  // id?: number;
-  charismaLevel: number;
-  sensibilityLevel: number;
-  insightLevel: number;
-  willingnessLevel: number;
-  handicraftLevel: number;
-  charmLevel: number;
-}
