+++ src/common/dto/propensity.dto.ts
@@ -1,28 +1,21 @@
 import { Expose } from 'class-transformer';
-import { IsNumber } from 'class-validator';
 
-export class CharacterPropensityDTO {
+export class PropensityDto {
   @Expose()
-  @IsNumber()
   charismaLevel: number;
 
   @Expose()
-  @IsNumber()
   sensibilityLevel: number;
 
   @Expose()
-  @IsNumber()
   insightLevel: number;
 
   @Expose()
-  @IsNumber()
   willingnessLevel: number;
 
   @Expose()
-  @IsNumber()
   handicraftLevel: number;
 
   @Expose()
-  @IsNumber()
   charmLevel: number;
 }
