+++ src/nxapi/nxapi.service.ts
@@ -1,6 +1,18 @@
 import { HttpService } from '@nestjs/axios';
 import { Injectable, Logger, OnModuleInit } from '@nestjs/common';
 import axiosRetry from 'axios-retry';
+import { AbilityDto } from 'src/common/dto/ability.dto';
+import { CharacterBasicDto } from 'src/common/dto/character-basic.dto';
+import { HyperStatPresetDto } from 'src/common/dto/hyper-stat.dto';
+import { ItemEquipmentPresetDto } from 'src/common/dto/item-equipment.dto';
+import { PropensityDto } from 'src/common/dto/propensity.dto';
+import { StatDto } from 'src/common/dto/stat.dto';
+import { abilityMapper } from './mapper/ability.mapper';
+import { characterBasicMapper } from './mapper/character-basic.mapper';
+import { hyperStatMapper } from './mapper/hyper-stat.mapper';
+import { itemEquipmentMapper } from './mapper/item-equipment.mapper';
+import { propensityMapper } from './mapper/propensity.mapper';
+import { statMapper } from './mapper/stat.mapper';
 import { NxapiAbilityData } from './type/nxapi-ability.type';
 import { NxapiItemEquipment } from './type/nxapi-item-equipment.type';
 
@@ -21,7 +33,7 @@ export class NxapiService implements OnModuleInit {
     });
   }
 
-  private async nxapi(url: string, params: { [key: string]: string }) {
+  private async nxapi<T>(url: string, params: { [key: string]: string }): Promise<T> {
     try {
       const res = await this.httpService.axiosRef.get(url, { params });
       return res.data;
@@ -32,163 +44,163 @@ export class NxapiService implements OnModuleInit {
   }
 
   async fetchCharacterOcid(characterName: string): Promise<string> {
-    const res = await this.nxapi('/id', { character_name: characterName });
+    const res = await this.nxapi<any>('/id', { character_name: characterName });
     this.logger.log(`fetchCharacterOcid: ${characterName} => ${res.ocid}`);
     return res.ocid;
   }
 
-  async fetchCharacterBasic(ocid: string, date?: string): Promise<object> {
-    const res = await this.nxapi('/character/basic', { ocid, date });
-    this.logger.log(`fetchCharacterBasic: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`);
-    return res;
+  async fetchCharacterBasic(ocid: string, date?: string): Promise<CharacterBasicDto> {
+    const res = await this.nxapi<any>('/character/basic', { ocid, date });
+    this.logger.log(`fetchCharacterBasic: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`);
+    return characterBasicMapper(ocid, res);
   }
 
-  async fetchCharacterPopularity(ocid: string, date?: string): Promise<object> {
-    const res = await this.nxapi('/character/popularity', { ocid, date });
+  async fetchCharacterPopularity(ocid: string, date?: string): Promise<number> {
+    const res = await this.nxapi<any>('/character/popularity', { ocid, date });
     this.logger.log(
-      `fetchCharacterPopularity: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`,
+      `fetchCharacterPopularity: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`,
     );
-    return res;
+    return res.popularity;
   }
 
-  async fetchCharacterStat(ocid: string, date?: string): Promise<object> {
-    const res = await this.nxapi('/character/stat', { ocid, date });
-    this.logger.log(`fetchCharacterStat: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`);
-    return res;
+  async fetchCharacterStat(ocid: string, date?: string): Promise<StatDto> {
+    const res = await this.nxapi<any>('/character/stat', { ocid, date });
+    this.logger.log(`fetchCharacterStat: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`);
+    return statMapper(res);
   }
 
-  async fetchCharacterHyperStat(ocid: string, date?: string): Promise<object> {
-    const res = await this.nxapi('/character/hyper-stat', { ocid, date });
+  async fetchCharacterHyperStat(ocid: string, date?: string): Promise<HyperStatPresetDto[]> {
+    const res = await this.nxapi<any>('/character/hyper-stat', { ocid, date });
     this.logger.log(
-      `fetchCharacterHyperStat: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`,
+      `fetchCharacterHyperStat: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`,
     );
-    return res;
+    return hyperStatMapper(res);
   }
 
-  async fetchCharacterPropensity(ocid: string, date?: string): Promise<object> {
-    const res = await this.nxapi('/character/propensity', { ocid, date });
+  async fetchCharacterPropensity(ocid: string, date?: string): Promise<PropensityDto> {
+    const res = await this.nxapi<any>('/character/propensity', { ocid, date });
     this.logger.log(
-      `fetchCharacterPropensity: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`,
+      `fetchCharacterPropensity: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`,
     );
-    return res;
+    return propensityMapper(res);
   }
 
-  async fetchCharacterAbility(ocid: string, date?: string): Promise<NxapiAbilityData> {
-    const res = await this.nxapi('/character/ability', { ocid, date });
+  async fetchCharacterAbility(ocid: string, date?: string): Promise<AbilityDto> {
+    const res = await this.nxapi<NxapiAbilityData>('/character/ability', { ocid, date });
     this.logger.log(
-      `fetchCharacterAbility: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`,
+      `fetchCharacterAbility: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`,
     );
-    return res;
+    return abilityMapper(res);
   }
 
-  async fetchCharacterItemEquipment(ocid: string, date?: string): Promise<NxapiItemEquipment> {
-    const res = await this.nxapi('/character/item-equipment', { ocid, date });
+  async fetchCharacterItemEquipment(ocid: string, date?: string): Promise<ItemEquipmentPresetDto[]> {
+    const res = await this.nxapi<NxapiItemEquipment>('/character/item-equipment', { ocid, date });
     this.logger.log(
-      `fetchCharacterItemEquipment: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`,
+      `fetchCharacterItemEquipment: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`,
     );
-    return res;
+    return itemEquipmentMapper(res);
   }
 
   async fetchCharacterCashitemEquipment(ocid: string, date?: string): Promise<object> {
-    const res = await this.nxapi('/character/cashitem-equipment', {
+    const res = await this.nxapi<any>('/character/cashitem-equipment', {
       ocid,
       date,
     });
     this.logger.log(
-      `fetchCharacterCashitemEquipment: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`,
+      `fetchCharacterCashitemEquipment: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`,
     );
     return res;
   }
 
   async fetchCharacterSymbolEquipment(ocid: string, date?: string): Promise<object> {
-    const res = await this.nxapi('/character/symbol-equipment', { ocid, date });
+    const res = await this.nxapi<any>('/character/symbol-equipment', { ocid, date });
     this.logger.log(
-      `fetchCharacterSymbolEquipment: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`,
+      `fetchCharacterSymbolEquipment: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`,
     );
     return res;
   }
 
   async fetchCharacterSetEffect(ocid: string, date?: string): Promise<object> {
-    const res = await this.nxapi('/character/set-effect', { ocid, date });
+    const res = await this.nxapi<any>('/character/set-effect', { ocid, date });
     this.logger.log(
-      `fetchCharacterSetEffect: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`,
+      `fetchCharacterSetEffect: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`,
     );
     return res;
   }
 
   async fetchCharacterBeautyEquipment(ocid: string, date?: string): Promise<object> {
-    const res = await this.nxapi('/character/beauty-equipment', { ocid, date });
+    const res = await this.nxapi<any>('/character/beauty-equipment', { ocid, date });
     this.logger.log(
-      `fetchCharacterBeautyEquipment: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`,
+      `fetchCharacterBeautyEquipment: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`,
     );
     return res;
   }
 
   async fetchCharacterAndroidEquipment(ocid: string, date?: string): Promise<object> {
-    const res = await this.nxapi('/character/android-equipment', {
+    const res = await this.nxapi<any>('/character/android-equipment', {
       ocid,
       date,
     });
     this.logger.log(
-      `fetchCharacterAndroidEquipment: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`,
+      `fetchCharacterAndroidEquipment: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`,
     );
     return res;
   }
 
   async fetchCharacterPetEquipment(ocid: string, date?: string): Promise<object> {
-    const res = await this.nxapi('/character/pet-equipment', { ocid, date });
+    const res = await this.nxapi<any>('/character/pet-equipment', { ocid, date });
     this.logger.log(
-      `fetchCharacterPetEquipment: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`,
+      `fetchCharacterPetEquipment: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`,
     );
     return res;
   }
 
   async fetchCharacterSkill(ocid: string, skillGrade: SkillGrade, date?: string): Promise<object> {
-    const res = await this.nxapi('/character/skill', {
+    const res = await this.nxapi<any>('/character/skill', {
       ocid,
       date,
       character_skill_grade: skillGrade,
     });
-    this.logger.log(`fetchCharacterSkill: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`);
+    this.logger.log(`fetchCharacterSkill: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`);
     return res;
   }
 
   async fetchCharacterLinkSkill(ocid: string, date?: string): Promise<object> {
-    const res = await this.nxapi('/character/link-skill', { ocid, date });
+    const res = await this.nxapi<any>('/character/link-skill', { ocid, date });
     this.logger.log(
-      `fetchCharacterLinkSkill: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`,
+      `fetchCharacterLinkSkill: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`,
     );
     return res;
   }
 
   async fetchCharacterVmatrix(ocid: string, date?: string): Promise<object> {
-    const res = await this.nxapi('/character/vmatrix', { ocid, date });
+    const res = await this.nxapi<any>('/character/vmatrix', { ocid, date });
     this.logger.log(
-      `fetchCharacterVmatrix: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`,
+      `fetchCharacterVmatrix: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`,
     );
     return res;
   }
 
   async fetchCharacterHexamatrix(ocid: string, date?: string): Promise<object> {
-    const res = await this.nxapi('/character/hexamatrix', { ocid, date });
+    const res = await this.nxapi<any>('/character/hexamatrix', { ocid, date });
     this.logger.log(
-      `fetchCharacterHexamatrix: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`,
+      `fetchCharacterHexamatrix: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`,
     );
     return res;
   }
 
   async fetchCharacterHexamatrixStat(ocid: string, date?: string): Promise<object> {
-    const res = await this.nxapi('/character/hexamatrix-stat', { ocid, date });
+    const res = await this.nxapi<any>('/character/hexamatrix-stat', { ocid, date });
     this.logger.log(
-      `fetchCharacterHexamatrixStat: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`,
+      `fetchCharacterHexamatrixStat: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`,
     );
     return res;
   }
 
   async fetchCharacterDojang(ocid: string, date?: string): Promise<object> {
-    const res = await this.nxapi('/character/dojang', { ocid, date });
+    const res = await this.nxapi<any>('/character/dojang', { ocid, date });
     this.logger.log(
-      `fetchCharacterDojang: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 30) + '...'}`,
+      `fetchCharacterDojang: ${ocid.slice(0, 10) + '...'} => ${JSON.stringify(res).slice(0, 50) + '...'}`,
     );
     return res;
   }
